v0.6a — Split monolith + Jinja2 templating
fetch_news.py (~2200 lines) split into focused modules with no logic changes:
- memory.py — all memory/health/cache functions
- api.py — Claude wrappers, log_api_call, relative_time, format_articles_for_prompt
- fetchers.py — all RSS/HTTP fetchers + fetch_world_topic_sources
- processors.py — all category processors, world topics, developing situations, HEADLINE_RULES
- fetch_news.py — entry point only (orchestration + mock_data)

HTML extracted from f-string into Jinja2 template:
- page/builder.py — build_html() renders via Jinja2 Environment, holds ACCENTS
- page/template.html — full HTML/CSS/JS with unified render_story macro

render_story() consolidation:
- Three diverged versions (column, breaking grid, yesterday breaking) unified into
  one Jinja2 macro with variant parameter ('column'|'breaking'|'yesterday_breaking')

Notes:
- Directory named page/ not html/ — html/ shadows Python stdlib html module
  (feedparser imports html.entities), breaking the import chain
- log_api_call lives in api.py — only called from API wrappers; memory.py
  placement would require api.py to import internally
- jinja2 added to requirements.txt

Syntax checks pass: import fetch_news OK, import modules OK, mock render 74k chars
