+++ requirements.txt
@@ -2,3 +2,4 @@ anthropic
 requests
 feedparser
 beautifulsoup4
+jinja2
